More fixes for selecting backends and abstracting.
