+++ islandora_solr.module
@@ -642,3 +642,30 @@ function islandora_solr_get_current_backend() {
 function islandora_solr_get_namespaced_class(array $backend, $class) {
   return $backend['namespace'] . '\\' . $backend[$class];
 }
+
+/**
+ * Filter backends missing fields.
+ *
+ * @param $backend
+ *   The query backend array.
+ * @return bool
+ *   Whether the backend definition is complete.
+ */
+function islandora_solr_filter_query_backends($backend) {
+  $expected_tags = array(
+    'title' => 1,
+    'namespace' => 1,
+    'queryClass' => 1,
+    'resultsClass' => 1,
+    'facetClass' => 1,
+    'minVersion' => 1,
+    'maxVersion' => 1
+  );
+  if (is_array($backend)) {
+    $diff = array_diff_key($expected_tags, $backend);
+    if (count($diff) == 0) {
+      return TRUE;
+    }
+  }
+  return FALSE;
+}
